+++ ufp/__init__.py
@@ -4,14 +4,14 @@
 from __future__ import unicode_literals, absolute_import
 
 __title__ = 'ufp'
-__version__ = '1.12'
+__version__ = '1.13'
 __author__ = '별님'
 __author_email__ = 'w7dn1ng75r@gmail.com'
 __license__ = 'GPL v3'
 __copyright__ = 'Copyright 2015 별님'
 __url__ = 'https://github.com/Thestars3/pyufp/'
 
-from .ufp import *
-
 #하위 호환성 유지를 위한 설정
 from . import _compatibleness
+
+from .ufp import *